+++ src/compiler/app/app-global-scripts.ts
@@ -1,8 +1,8 @@
 import { AppRegistry, BuildCtx, CompilerCtx, Config, SourceTarget } from '../../declarations';
-import { buildExpressionReplacer } from '../build/replacer';
 import { createOnWarnFn, loadRollupDiagnostics } from '../../util/logger/logger-rollup';
 import { generatePreamble } from '../util';
 import { getGlobalEsmBuildPath, getGlobalFileName, getGlobalJsBuildPath } from './app-file-naming';
+import rollupPluginReplace from '../bundle/rollup-plugins/rollup-plugin-replace';
 import inMemoryFsRead from '../bundle/rollup-plugins/in-memory-fs-read';
 import { minifyJs } from '../minifier';
 import { transpileToEs5Main } from '../transpile/transpile-to-es5-main';
@@ -61,9 +61,8 @@ export async function generateAppGlobalContent(config: Config, compilerCtx: Comp
 async function loadDependentGlobalJsContents(config: Config, compilerCtx: CompilerCtx, buildCtx: BuildCtx, sourceTarget: SourceTarget) {
   const collections = compilerCtx.collections.filter(m => m.global && m.global.jsFilePath);
 
-  const dependentGlobalJsContents = await Promise.all(collections.map(async collectionManifest => {
-    const dependentGlobalJsContent = await bundleProjectGlobal(config, compilerCtx, buildCtx, sourceTarget, collectionManifest.collectionName, collectionManifest.global.jsFilePath);
-    return dependentGlobalJsContent;
+  const dependentGlobalJsContents = await Promise.all(collections.map(collectionManifest => {
+    return bundleProjectGlobal(config, compilerCtx, buildCtx, sourceTarget, collectionManifest.collectionName, collectionManifest.global.jsFilePath);
   }));
 
   return dependentGlobalJsContents;
@@ -85,9 +84,10 @@ async function bundleProjectGlobal(config: Config, compilerCtx: CompilerCtx, bui
   // ok, so the project also provided an entry file, so let's bundle it up and
   // the output from this can be tacked onto the top of the project's core file
   // start the bundler on our temporary file
-  let output = '';
-
   try {
+    const replaceObj = {
+      'process.env.NODE_ENV': config.devMode ? '"development"' : '"production"'
+    };
     const rollup = await config.sys.rollup.rollup({
       input: entry,
       plugins: [
@@ -100,6 +100,9 @@ async function bundleProjectGlobal(config: Config, compilerCtx: CompilerCtx, bui
           include: 'node_modules/**',
           sourceMap: false
         }),
+        rollupPluginReplace({
+          values: replaceObj
+        }),
         inMemoryFsRead(config, compilerCtx, buildCtx),
         ...config.plugins
       ],
@@ -109,21 +112,16 @@ async function bundleProjectGlobal(config: Config, compilerCtx: CompilerCtx, bui
     const results = await rollup.generate({ format: 'es' });
 
     // cool, so we balled up all of the globals into one string
-
-    // replace build time expressions, like process.env.NODE_ENV === 'production'
-    // with a hard coded boolean
-    results.code = buildExpressionReplacer(config, results.code);
+    buildCtx.global = compilerCtx.moduleFiles[config.globalScript];
 
     // wrap our globals code with our own iife
-    output = await wrapGlobalJs(config, compilerCtx, buildCtx, sourceTarget, namespace, results.code);
-
-    buildCtx.global = compilerCtx.moduleFiles[config.globalScript];
+    return await wrapGlobalJs(config, compilerCtx, buildCtx, sourceTarget, namespace, results.code);
 
   } catch (e) {
     loadRollupDiagnostics(config, compilerCtx, buildCtx, e);
   }
 
-  return output;
+  return '';
 }
 
 
@@ -133,13 +131,10 @@ async function wrapGlobalJs(config: Config, compilerCtx: CompilerCtx, buildCtx: 
   // just format it a touch better in dev mode
   jsContent = `\n/** ${globalJsName || ''} global **/\n\n${jsContent}`;
 
-  const lines = jsContent.split(/\r?\n/);
-  jsContent = lines.map(line => {
-    if (line.length) {
-      return '    ' + line;
-    }
-    return line;
-  }).join('\n');
+  jsContent = jsContent
+    .split(/\r?\n/)
+    .map(line => (line.length > 0) ? '    ' + line : line)
+    .join('\n');
 
   if (sourceTarget === 'es5') {
     // global could already be in es2017