+++ src/compiler/bundle/rollup-bundle.ts
@@ -6,7 +6,6 @@ import { getUserCompilerOptions } from '../transpile/compiler-options';
 import localResolution from './rollup-plugins/local-resolution';
 import inMemoryFsRead from './rollup-plugins/in-memory-fs-read';
 import { RollupBuild, RollupDirOptions } from 'rollup'; // types only
-import nodeEnvVars from './rollup-plugins/node-env-vars';
 import pathsResolution from './rollup-plugins/paths-resolution';
 import pluginHelper from './rollup-plugins/plugin-helper';
 import rollupPluginReplace from './rollup-plugins/rollup-plugin-replace';
@@ -18,19 +17,12 @@ export async function createBundle(config: d.Config, compilerCtx: d.CompilerCtx,
     buildCtx.debug(`createBundle aborted, not active build`);
   }
 
-  const buildConditionals = {
-    isDev: !!config.devMode
-  } as d.BuildConditionals;
-
-  const replaceObj = Object.keys(buildConditionals).reduce((all, key) => {
-    all[`Build.${key}`] = buildConditionals[key];
-    return all;
-  }, <{ [key: string]: any}>{});
-
   const timeSpan = buildCtx.createTimeSpan(`createBundle started`, true);
 
-  let rollupBundle: RollupBuild;
-
+  const replaceObj = {
+    'Build.isDev': !!config.devMode,
+    'process.env.NODE_ENV': config.devMode ? '"development"' : '"production"'
+  };
   const commonjsConfig = {
     include: 'node_modules/**',
     sourceMap: false,
@@ -54,9 +46,6 @@ export async function createBundle(config: d.Config, compilerCtx: d.CompilerCtx,
     cache: config.enableCache ? compilerCtx.rollupCache : undefined,
     plugins: [
       abortPlugin(buildCtx),
-      rollupPluginReplace({
-        values: replaceObj
-      }),
       config.sys.rollup.plugins.nodeResolve(nodeResolveConfig),
       config.sys.rollup.plugins.emptyJsResolver(),
       config.sys.rollup.plugins.commonjs(commonjsConfig),
@@ -64,7 +53,9 @@ export async function createBundle(config: d.Config, compilerCtx: d.CompilerCtx,
       inMemoryFsRead(config, compilerCtx, buildCtx, entryModules),
       pathsResolution(config, compilerCtx, tsCompilerOptions),
       localResolution(config, compilerCtx),
-      nodeEnvVars(config),
+      rollupPluginReplace({
+        values: replaceObj
+      }),
       ...config.plugins,
       statsPlugin(buildCtx),
       pluginHelper(config, compilerCtx, buildCtx),
@@ -73,8 +64,11 @@ export async function createBundle(config: d.Config, compilerCtx: d.CompilerCtx,
     onwarn: createOnWarnFn(config, buildCtx.diagnostics)
   };
 
+  let rollupBundle: RollupBuild;
   try {
     rollupBundle = await config.sys.rollup.rollup(rollupConfig);
+    compilerCtx.rollupCache = rollupBundle ? rollupBundle.cache : undefined;
+
   } catch (err) {
     // clean rollup cache if error
     compilerCtx.rollupCache = undefined;
@@ -87,9 +81,7 @@ export async function createBundle(config: d.Config, compilerCtx: d.CompilerCtx,
       buildCtx.debug(`createBundle errors ignored, not active build`);
     }
   }
-  compilerCtx.rollupCache = rollupBundle.cache;
 
   timeSpan.finish(`createBundle finished`);
-
   return rollupBundle;
 }