+++ src/compiler/distribution/dist-angular.ts
@@ -1,6 +1,8 @@
 import * as d from '../../declarations';
 import { dashToPascalCase } from '../../util/helpers';
 import { MEMBER_TYPE } from '../../util/constants';
+import { isDocsPublic } from '../util';
+import { OutputTargetAngular } from '../../declarations';
 
 
 export async function generateAngularProxies(config: d.Config, compilerCtx: d.CompilerCtx, cmpRegistry: d.ComponentRegistry) {
@@ -12,11 +14,22 @@ export async function generateAngularProxies(config: d.Config, compilerCtx: d.Co
   }));
 }
 
+async function angularDirectiveProxyOutput(config: d.Config, compilerCtx: d.CompilerCtx, outputTarget: d.OutputTargetAngular, cmpRegistry: d.ComponentRegistry) {
+  const components = getComponents(outputTarget.excludeComponents, cmpRegistry);
+
+  await Promise.all([
+    generateProxies(config, compilerCtx, components, outputTarget),
+    generateAngularArray(config, compilerCtx, components, outputTarget),
+    generateAngularUtils(compilerCtx, outputTarget)
+  ]);
+
+  config.logger.debug(`generated angular directives: ${outputTarget.directivesProxyFile}`);
+}
 
 function getComponents(excludeComponents: string[], cmpRegistry: d.ComponentRegistry): d.ComponentMeta[] {
   return Object.keys(cmpRegistry)
     .map(key => cmpRegistry[key])
-    .filter(c => !excludeComponents.includes(c.tagNameMeta))
+    .filter(c => !excludeComponents.includes(c.tagNameMeta) && isDocsPublic(c.jsdoc))
     .sort((a, b) => {
       if (a.tagNameMeta < b.tagNameMeta) return -1;
       if (a.tagNameMeta > b.tagNameMeta) return 1;
@@ -24,127 +37,34 @@ function getComponents(excludeComponents: string[], cmpRegistry: d.ComponentRegi
     });
 }
 
-async function angularDirectiveProxyOutput(config: d.Config, compilerCtx: d.CompilerCtx, outputTarget: d.OutputTargetAngular, cmpRegistry: d.ComponentRegistry) {
-  const components = getComponents(outputTarget.excludeComponents, cmpRegistry);
-  const useDirectives = outputTarget.useDirectives;
-  const { hasOutputs, proxies } = generateProxies(components, useDirectives);
+async function generateProxies(config: d.Config, compilerCtx: d.CompilerCtx, components: d.ComponentMeta[], outputTarget: OutputTargetAngular) {
+  const proxies = getProxies(components);
 
-  const auxFunctions: string[] = [
-    inputsAuxFunction(),
-    outputsAuxFunction(),
-    methodsAuxFunction()
-  ];
-  const angularImports = [
-    'ElementRef'
-  ];
-
-  if (components.length > 0) {
-    if (useDirectives) {
-      angularImports.push('Directive');
-    } else {
-      angularImports.push('Component');
-      angularImports.push('ViewEncapsulation');
-      angularImports.push('ChangeDetectionStrategy');
-      angularImports.push('ChangeDetectorRef');
-    }
-  }
-
-  if (hasOutputs) {
-    angularImports.push('EventEmitter');
-  }
-
-  const imports = `
-${hasOutputs ? `import { fromEvent } from 'rxjs';` : '' }
-import { ${angularImports.sort().join(', ')} } from '@angular/core';
-`;
+  const imports = `/* tslint:disable */
+/* auto-generated angular directive proxies */
+import { Component, ElementRef, ChangeDetectorRef, EventEmitter } from '@angular/core';`;
 
   const sourceImports = !outputTarget.componentCorePackage ? ''
     : `type StencilComponents<T extends keyof StencilElementInterfaces> = StencilElementInterfaces[T];`;
 
   const final: string[] = [
-    '/* tslint:disable */',
-    '/* auto-generated angular directive proxies */',
     imports,
+    getProxyUtils(config, outputTarget),
     sourceImports,
-    auxFunctions.join('\n'),
     proxies,
   ];
 
   const finalText = final.join('\n') + '\n';
   await compilerCtx.fs.writeFile(outputTarget.directivesProxyFile, finalText);
-  if (outputTarget.directivesArrayFile) {
-    const proxyPath = relativeImport(config, outputTarget.directivesArrayFile, outputTarget.directivesProxyFile);
-    const a = angularArray(components, proxyPath);
-    await compilerCtx.fs.writeFile(outputTarget.directivesArrayFile, a);
-  }
-  config.logger.debug(`generated angular directives: ${outputTarget.directivesProxyFile}`);
 }
 
-function inputsAuxFunction() {
-  return `
-export function proxyInputs(instance: any, el: any, props: string[]) {
-  props.forEach(propName => {
-    Object.defineProperty(instance, propName, {
-      get: () => el[propName], set: (val: any) => el[propName] = val
-    });
-  });
-}`;
+function getProxies(components: d.ComponentMeta[]) {
+  return components
+    .map(getProxy)
+    .join('\n');
 }
 
-
-function outputsAuxFunction() {
-  return `
-export function proxyOutputs(instance: any, el: any, events: string[]) {
-  events.forEach(eventName => instance[eventName] = fromEvent(el, eventName));
-}`;
-}
-
-
-function methodsAuxFunction() {
-  return `
-export function proxyMethods(instance: any, el: any, methods: string[]) {
-  methods.forEach(methodName => {
-    Object.defineProperty(instance, methodName, {
-      get: function() {
-        return function() {
-          const args = arguments;
-          return el.componentOnReady().then((el: any) => el[methodName].apply(el, args));
-        };
-      }
-    });
-  });
-}
-`;
-}
-
-function generateProxies(components: d.ComponentMeta[], useDirectives: boolean) {
-  let hasMethods = false;
-  let hasOutputs = false;
-  let hasInputs = false;
-
-  const lines = components.map(cmpMeta => {
-    const proxy = generateProxy(cmpMeta, useDirectives);
-    if (proxy.hasInputs) {
-      hasInputs = true;
-    }
-    if (proxy.hasMethods) {
-      hasMethods = true;
-    }
-    if (proxy.hasOutputs) {
-      hasOutputs = true;
-    }
-    return proxy.text;
-  });
-
-  return {
-    proxies: lines.join('\n'),
-    hasInputs,
-    hasMethods,
-    hasOutputs
-  };
-}
-
-function generateProxy(cmpMeta: d.ComponentMeta, useDirectives: boolean) {
+function getProxy(cmpMeta: d.ComponentMeta) {
   // Collect component meta
   const inputs = getInputs(cmpMeta);
   const outputs = getOutputs(cmpMeta);
@@ -154,20 +74,13 @@ function generateProxy(cmpMeta: d.ComponentMeta, useDirectives: boolean) {
   const hasInputs = inputs.length > 0;
   const hasOutputs = outputs.length > 0;
   const hasMethods = methods.length > 0;
-  const hasContructor = hasInputs || hasOutputs || hasMethods;
 
   // Generate Angular @Directive
-  const decorator = useDirectives ? 'Directive' : 'Component';
   const directiveOpts = [
     `selector: \'${cmpMeta.tagNameMeta}\'`,
+    `changeDetection: 0`,
+    `template: '<ng-content></ng-content>'`
   ];
-  if (!useDirectives) {
-    directiveOpts.push(
-      `changeDetection: ChangeDetectionStrategy.OnPush`,
-      `encapsulation: ViewEncapsulation.None`,
-      `template: '<ng-content></ng-content>'`
-    );
-  }
   if (inputs.length > 0) {
     directiveOpts.push(`inputs: ['${inputs.join(`', '`)}']`);
   }
@@ -175,7 +88,7 @@ function generateProxy(cmpMeta: d.ComponentMeta, useDirectives: boolean) {
   const tagNameAsPascal = dashToPascalCase(cmpMeta.tagNameMeta);
   const lines = [`
 export declare interface ${tagNameAsPascal} extends StencilComponents<'${tagNameAsPascal}'> {}
-@${decorator}({ ${directiveOpts.join(', ')} })
+@Component({ ${directiveOpts.join(', ')} })
 export class ${tagNameAsPascal} {`];
 
   // Generate outputs
@@ -183,64 +96,81 @@ export class ${tagNameAsPascal} {`];
     lines.push(`  ${output}!: EventEmitter<CustomEvent>;`);
   });
 
-  // Generate component constructor
-  if (hasContructor) {
-    if (useDirectives) {
-      lines.push(`
-  constructor(r: ElementRef) {
-    const el = r.nativeElement;`);
-    } else {
-      lines.push(`
-  constructor(c: ChangeDetectorRef, r: ElementRef) {
+  lines.push('  el: HTMLElement;');
+  lines.push(`  constructor(c: ChangeDetectorRef, r: ElementRef) {
     c.detach();
-    const el = r.nativeElement;`);
-    }
-  }
-
-  if (hasMethods) {
-    lines.push(`    proxyMethods(this, el, ['${methods.join(`', '`)}']);`);
-  }
-
-  if (hasInputs) {
-    lines.push(`    proxyInputs(this, el, ['${inputs.join(`', '`)}']);`);
-  }
-
+    this.el = r.nativeElement;`);
   if (hasOutputs) {
-    lines.push(`    proxyOutputs(this, el, ['${outputs.join(`', '`)}']);`);
+    lines.push(`    proxyOutputs(this, this.el, ['${outputs.join(`', '`)}']);`);
   }
-
-  if (hasContructor) {
-    lines.push(`  }`);
-  }
-
+  lines.push(`  }`);
   lines.push(`}`);
 
-  return {
-    text: lines.join('\n'),
-    hasInputs,
-    hasMethods,
-    hasOutputs
-  };
+  if (hasMethods) {
+    lines.push(`proxyMethods(${tagNameAsPascal}, ['${methods.join(`', '`)}']);`);
+  }
+  if (hasInputs) {
+    lines.push(`proxyInputs(${tagNameAsPascal}, ['${inputs.join(`', '`)}']);`);
+  }
+
+  return lines.join('\n');
 }
 
 function getInputs(cmpMeta: d.ComponentMeta) {
   return Object.keys(cmpMeta.membersMeta || {}).filter(memberName => {
     const m = cmpMeta.membersMeta[memberName];
-    return m.memberType === MEMBER_TYPE.Prop || m.memberType === MEMBER_TYPE.PropMutable;
+    return isDocsPublic(m.jsdoc) && (m.memberType === MEMBER_TYPE.Prop || m.memberType === MEMBER_TYPE.PropMutable);
   });
 }
 
 function getOutputs(cmpMeta: d.ComponentMeta) {
-  return (cmpMeta.eventsMeta || []).map(eventMeta => eventMeta.eventName);
+  return (cmpMeta.eventsMeta || [])
+    .filter(e => isDocsPublic(e.jsdoc))
+    .map(eventMeta => eventMeta.eventName);
 }
 
 function getMethods(cmpMeta: d.ComponentMeta) {
   return Object.keys(cmpMeta.membersMeta || {}).filter(memberName => {
     const m = cmpMeta.membersMeta[memberName];
-    return m.memberType === MEMBER_TYPE.Method;
+    return isDocsPublic(m.jsdoc) && m.memberType === MEMBER_TYPE.Method;
   });
 }
 
+function getProxyUtils(config: d.Config, outputTarget: OutputTargetAngular) {
+  if (!outputTarget.directivesUtilsFile) {
+    return PROXY_UTILS.replace(/export function/g, 'function');
+  } else {
+    const utilsPath = relativeImport(config, outputTarget.directivesProxyFile, outputTarget.directivesUtilsFile);
+    return `import { proxyInputs, proxyMethods, proxyOutputs } from '${utilsPath}';\n`;
+  }
+}
+
+async function generateAngularArray(config: d.Config, compilerCtx: d.CompilerCtx, components: d.ComponentMeta[], outputTarget: OutputTargetAngular) {
+  if (!outputTarget.directivesArrayFile) {
+    return;
+  }
+
+  const proxyPath = relativeImport(config, outputTarget.directivesArrayFile, outputTarget.directivesProxyFile);
+  const directives = components
+    .map(cmpMeta => dashToPascalCase(cmpMeta.tagNameMeta))
+    .map(className => `d.${className}`)
+    .join(',\n  ');
+
+  const c = `
+import * as d from '${proxyPath}';
+
+export const DIRECTIVES = [
+${directives}
+];
+`;
+  await compilerCtx.fs.writeFile(outputTarget.directivesArrayFile, c);
+}
+
+async function generateAngularUtils(compilerCtx: d.CompilerCtx, outputTarget: OutputTargetAngular) {
+  if (outputTarget.directivesUtilsFile) {
+    await compilerCtx.fs.writeFile(outputTarget.directivesUtilsFile, '/* tslint:disable */\n' + PROXY_UTILS);
+  }
+}
 
 function relativeImport(config: d.Config, pathFrom: string, pathTo: string) {
   let relativePath = config.sys.path.relative(config.sys.path.dirname(pathFrom), config.sys.path.dirname(pathTo));
@@ -248,17 +178,30 @@ function relativeImport(config: d.Config, pathFrom: string, pathTo: string) {
   return `${relativePath}/${config.sys.path.basename(pathTo, '.ts')}`;
 }
 
-function angularArray(components: d.ComponentMeta[], proxyPath: string) {
-  const directives = components
-    .map(cmpMeta => dashToPascalCase(cmpMeta.tagNameMeta))
-    .map(className => `d.${className}`)
-    .join(',\n  ');
+const PROXY_UTILS = `import { fromEvent } from 'rxjs';
 
-  return `
-import * as d from '${proxyPath}';
-
-export const DIRECTIVES = [
-  ${directives}
-];
-`;
+export function proxyInputs(Cmp: any, inputs: string[]) {
+  const Prototype = Cmp.prototype;
+  inputs.forEach(item => {
+    Object.defineProperty(Prototype, item, {
+      get() { return this.el[item]; },
+      set(val: any) { this.el[item] = val; },
+    });
+  });
 }
+
+export function proxyMethods(Cmp: any, methods: string[]) {
+  const Prototype = Cmp.prototype;
+  methods.forEach(methodName => {
+    Prototype[methodName] = function() {
+      const args = arguments;
+      return this.el.componentOnReady().then((el: any) => el[methodName].apply(el, args));
+    };
+  });
+}
+
+export function proxyOutputs(instance: any, el: any, events: string[]) {
+  events.forEach(eventName => instance[eventName] = fromEvent(el, eventName));
+}
+`;
+