+++ src/compiler/distribution/dist-esm.ts
@@ -130,7 +130,10 @@ async function generateEsmLoader(config: d.Config, compilerCtx: d.CompilerCtx, o
     'name': 'loader',
     'typings': './index.d.ts',
     'module': './index.js',
-    'es2017': './index.es2017.js'
+    'jsnext:main': './index.es2017.js',
+    'es2015': './index.es2017.js',
+    'es2017': './index.es2017.js',
+    'sideEffects': false
   }, null, 2);
 
   const indexPath = config.buildEs5 ? es5EntryPoint : es2017EntryPoint;