+++ test/karma/test-app/components.d.ts
@@ -114,6 +114,15 @@ export namespace Components {
     'propVal'?: number;
   }
 
+  interface ExternalImportA {}
+  interface ExternalImportAAttributes extends StencilHTMLAttributes {}
+
+  interface ExternalImportB {}
+  interface ExternalImportBAttributes extends StencilHTMLAttributes {}
+
+  interface ExternalImportC {}
+  interface ExternalImportCAttributes extends StencilHTMLAttributes {}
+
   interface InitCssRoot {}
   interface InitCssRootAttributes extends StencilHTMLAttributes {}
 
@@ -194,6 +203,9 @@ export namespace Components {
 
   interface LifecycleUnloadRoot {}
   interface LifecycleUnloadRootAttributes extends StencilHTMLAttributes {}
+
+  interface ListenWindow {}
+  interface ListenWindowAttributes extends StencilHTMLAttributes {}
 
   interface NodeGlobals {}
   interface NodeGlobalsAttributes extends StencilHTMLAttributes {}
@@ -385,6 +397,9 @@ declare global {
     'DynamicImport': Components.DynamicImport;
     'Es5AddclassSvg': Components.Es5AddclassSvg;
     'EsmImport': Components.EsmImport;
+    'ExternalImportA': Components.ExternalImportA;
+    'ExternalImportB': Components.ExternalImportB;
+    'ExternalImportC': Components.ExternalImportC;
     'InitCssRoot': Components.InitCssRoot;
     'KeyReorderRoot': Components.KeyReorderRoot;
     'KeyReorder': Components.KeyReorder;
@@ -400,6 +415,7 @@ declare global {
     'LifecycleUnloadA': Components.LifecycleUnloadA;
     'LifecycleUnloadB': Components.LifecycleUnloadB;
     'LifecycleUnloadRoot': Components.LifecycleUnloadRoot;
+    'ListenWindow': Components.ListenWindow;
     'NodeGlobals': Components.NodeGlobals;
     'ReflectToAttr': Components.ReflectToAttr;
     'ScopedBasicRoot': Components.ScopedBasicRoot;
@@ -460,6 +476,9 @@ declare global {
     'dynamic-import': Components.DynamicImportAttributes;
     'es5-addclass-svg': Components.Es5AddclassSvgAttributes;
     'esm-import': Components.EsmImportAttributes;
+    'external-import-a': Components.ExternalImportAAttributes;
+    'external-import-b': Components.ExternalImportBAttributes;
+    'external-import-c': Components.ExternalImportCAttributes;
     'init-css-root': Components.InitCssRootAttributes;
     'key-reorder-root': Components.KeyReorderRootAttributes;
     'key-reorder': Components.KeyReorderAttributes;
@@ -475,6 +494,7 @@ declare global {
     'lifecycle-unload-a': Components.LifecycleUnloadAAttributes;
     'lifecycle-unload-b': Components.LifecycleUnloadBAttributes;
     'lifecycle-unload-root': Components.LifecycleUnloadRootAttributes;
+    'listen-window': Components.ListenWindowAttributes;
     'node-globals': Components.NodeGlobalsAttributes;
     'reflect-to-attr': Components.ReflectToAttrAttributes;
     'scoped-basic-root': Components.ScopedBasicRootAttributes;
@@ -620,6 +640,24 @@ declare global {
     new (): HTMLEsmImportElement;
   };
 
+  interface HTMLExternalImportAElement extends Components.ExternalImportA, HTMLStencilElement {}
+  var HTMLExternalImportAElement: {
+    prototype: HTMLExternalImportAElement;
+    new (): HTMLExternalImportAElement;
+  };
+
+  interface HTMLExternalImportBElement extends Components.ExternalImportB, HTMLStencilElement {}
+  var HTMLExternalImportBElement: {
+    prototype: HTMLExternalImportBElement;
+    new (): HTMLExternalImportBElement;
+  };
+
+  interface HTMLExternalImportCElement extends Components.ExternalImportC, HTMLStencilElement {}
+  var HTMLExternalImportCElement: {
+    prototype: HTMLExternalImportCElement;
+    new (): HTMLExternalImportCElement;
+  };
+
   interface HTMLInitCssRootElement extends Components.InitCssRoot, HTMLStencilElement {}
   var HTMLInitCssRootElement: {
     prototype: HTMLInitCssRootElement;
@@ -708,6 +746,12 @@ declare global {
   var HTMLLifecycleUnloadRootElement: {
     prototype: HTMLLifecycleUnloadRootElement;
     new (): HTMLLifecycleUnloadRootElement;
+  };
+
+  interface HTMLListenWindowElement extends Components.ListenWindow, HTMLStencilElement {}
+  var HTMLListenWindowElement: {
+    prototype: HTMLListenWindowElement;
+    new (): HTMLListenWindowElement;
   };
 
   interface HTMLNodeGlobalsElement extends Components.NodeGlobals, HTMLStencilElement {}
@@ -968,6 +1012,9 @@ declare global {
     'dynamic-import': HTMLDynamicImportElement
     'es5-addclass-svg': HTMLEs5AddclassSvgElement
     'esm-import': HTMLEsmImportElement
+    'external-import-a': HTMLExternalImportAElement
+    'external-import-b': HTMLExternalImportBElement
+    'external-import-c': HTMLExternalImportCElement
     'init-css-root': HTMLInitCssRootElement
     'key-reorder-root': HTMLKeyReorderRootElement
     'key-reorder': HTMLKeyReorderElement
@@ -983,6 +1030,7 @@ declare global {
     'lifecycle-unload-a': HTMLLifecycleUnloadAElement
     'lifecycle-unload-b': HTMLLifecycleUnloadBElement
     'lifecycle-unload-root': HTMLLifecycleUnloadRootElement
+    'listen-window': HTMLListenWindowElement
     'node-globals': HTMLNodeGlobalsElement
     'reflect-to-attr': HTMLReflectToAttrElement
     'scoped-basic-root': HTMLScopedBasicRootElement
@@ -1043,6 +1091,9 @@ declare global {
     'dynamic-import': HTMLDynamicImportElement;
     'es5-addclass-svg': HTMLEs5AddclassSvgElement;
     'esm-import': HTMLEsmImportElement;
+    'external-import-a': HTMLExternalImportAElement;
+    'external-import-b': HTMLExternalImportBElement;
+    'external-import-c': HTMLExternalImportCElement;
     'init-css-root': HTMLInitCssRootElement;
     'key-reorder-root': HTMLKeyReorderRootElement;
     'key-reorder': HTMLKeyReorderElement;
@@ -1058,6 +1109,7 @@ declare global {
     'lifecycle-unload-a': HTMLLifecycleUnloadAElement;
     'lifecycle-unload-b': HTMLLifecycleUnloadBElement;
     'lifecycle-unload-root': HTMLLifecycleUnloadRootElement;
+    'listen-window': HTMLListenWindowElement;
     'node-globals': HTMLNodeGlobalsElement;
     'reflect-to-attr': HTMLReflectToAttrElement;
     'scoped-basic-root': HTMLScopedBasicRootElement;